+++ src/plugins/fmp/multicafeTools/XLSExportMoveResults.java
@@ -111,8 +111,8 @@ public class XLSExportMoveResults  extends XLSExport {
 			expAll.fileTimeImageLast = expi.fileTimeImageLast;
 			expi = expi.nextExperiment;
 		}
-		expAll.fileTimeImageFirstMinute = expAll.fileTimeImageFirst.toMillis()/60000;
-		expAll.fileTimeImageLastMinute = expAll.fileTimeImageLast.toMillis()/60000;
+		expAll.fileTimeImageFirstMinute = (long) (expAll.fileTimeImageFirst.toMillis()/60000d);
+		expAll.fileTimeImageLastMinute = (long) (expAll.fileTimeImageLast.toMillis()/60000d);
 		int nFrames = (int) ((expAll.fileTimeImageLastMinute - expAll.fileTimeImageFirstMinute)/expAll.getCagesFrameStep() +1) ;
 		if (expAll.getCagesFrameEnd() < nFrames) {
 			expAll.setCagesFrameEnd(nFrames-1);
